Refactor model runner to split by materialization
