+++ dbt_dry_run/node_runner/__init__.py
@@ -14,8 +14,6 @@ class NodeRunner(metaclass=ABCMeta):
     Runs a node and returns a result
     """
 
-    resource_type: Tuple[str, ...]
-
     def __init__(self, sql_runner: SQLRunner, results: Results):
         self._sql_runner = sql_runner
         self._results = results
@@ -42,15 +40,3 @@ class NodeRunner(metaclass=ABCMeta):
                 )
         else:
             return None
-
-
-def _get_runner_resource_types(
-    runner: Type[NodeRunner],
-) -> List[Tuple[str, Type[NodeRunner]]]:
-    return [(resource_type, runner) for resource_type in runner.resource_type]
-
-
-def get_runner_map(runners: List[Type[NodeRunner]]) -> Dict[str, Type[NodeRunner]]:
-    return dict(
-        itertools.chain(*[_get_runner_resource_types(runner) for runner in runners])
-    )
