+++ dbt_dry_run/node_runner/seed_runner.py
@@ -11,8 +11,6 @@ from dbt_dry_run.results import DryRunResult, DryRunStatus
 
 
 class SeedRunner(NodeRunner):
-    resource_type = ("seed",)
-
     DEFAULT_TYPE = BigQueryFieldType.STRING
     TYPE_MAP = {
         data_types.Text: BigQueryFieldType.STRING,
